+++ scripts/check_release_ready.sh
@@ -28,6 +28,7 @@ matlab_available() {
 run_step "gallery outputs" ./scripts/check_gallery_outputs.sh
 run_step "gallery metadata consistency" ./scripts/check_gallery_consistency.sh
 run_step "version metadata consistency" ./scripts/check_version_consistency.sh
+run_step "citation metadata" ./scripts/check_citation.sh
 run_step "documentation links" ./scripts/check_docs_links.sh
 run_step "API reference coverage" ./scripts/check_api_reference.sh
 run_step "MATLAB help text" ./scripts/check_matlab_help.sh